requirements.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.2
cssselect==1.2.0
decorator==5.1.1
Django==4.1
django-environ==0.10.0
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.14.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.0
jmespath==1.0.1
lxml==4.9.3
matplotlib-inline==0.1.6
packaging==23.1
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.39
Protego==0.2.1
psycopg2==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.15.1
pyOpenSSL==23.2.0
pytz==2023.3
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
scrapy-djangoitem==1.1.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tldextract==3.4.4
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
w3lib==2.1.1
wcwidth==0.2.6
zope.interface==6.0
